restructure, clean up, ahead and behind count ui
