+++ web/renderer/components/pageComponents/DatabasePage/ForRemotes/RemotesPage/FetchRemoteModal/RemoteBranchRow.tsx
@@ -1,0 +1,56 @@
+import { QueryHandler } from "@dolthub/react-components";
+import {
+  BranchFragment,
+  RemoteFragment,
+  useMergeBaseQuery,
+} from "@gen/graphql-types";
+import { OptionalRefParams } from "@lib/params";
+import { getBranchName } from "./utils";
+import SyncButton from "./SyncButton";
+import BehindAheadCount from "./BehindAheadCount";
+
+type Props = {
+  branch: BranchFragment;
+  remote: RemoteFragment;
+  currentBranch: string;
+  params: OptionalRefParams;
+};
+
+export default function RemoteBranchRow({
+  branch,
+  remote,
+  params,
+  currentBranch,
+}: Props) {
+  const { remoteAndBranchName, remoteBranchName } = getBranchName(
+    branch.branchName,
+  );
+  const res = useMergeBaseQuery({
+    variables: {
+      databaseName: params.databaseName,
+      branchName: currentBranch,
+      anotherBranch: remoteAndBranchName,
+    },
+  });
+
+  return (
+    <QueryHandler
+      result={{ ...res, data: res.data?.mergeBase }}
+      render={data => (
+        <tr>
+          <td>{remoteAndBranchName}</td>
+          <BehindAheadCount numbers={data} />
+          <td>
+            <SyncButton
+              numbers={data}
+              params={params}
+              remote={remote}
+              remoteBranchName={remoteBranchName}
+              currentBranch={currentBranch}
+            />
+          </td>
+        </tr>
+      )}
+    />
+  );
+}
